+++ deploy/restart-sequencer.sh
@@ -1,4 +1,12 @@
 #!/bin/sh
+if [ -z "$1" ]; then
+    echo "Error: No `branch` provided."
+    echo "Usage: $0 main"
+    echo "   or: $0 dev"
+    exit 1
+fi
+
+export SOCIALCAP_HOME=socialcap-$1
 
 # will run the Socialcap Sequencer with NO port
 sudo docker rm $(sudo docker stop sc-sequencer)
@@ -7,5 +15,5 @@ sudo docker -l debug run -d --restart=always --name sc-sequencer \
   --env MAIN=main-sequencer \
   --user $(id -u www-data):$(id -g www-data) \
   -v /etc/localtime:/etc/localtime:ro \
-  -v /home/socialcap/var:/var \
+  -v /home/SOCIALCAP_HOME/var:/var \
   socialcap/services:run
